Fixes failures related to stats collection w.r.t. long emulation across JSNI boundaries.
Patch by: jaimeyap
Review by: me


git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@2155 8db76d5a-ed1c-0410-87a9-c151d255dfc7

+++ user/src/com/google/gwt/user/client/rpc/impl/RequestCallbackAdapter.java
@@ -143,7 +143,7 @@ public class RequestCallbackAdapter<T> implements RequestCallback {
   /**
    * Used for stats recording.
    */
-  private final long requestId;
+  private final int requestId;
 
   /**
    * Instance which will read the expected return type out of the
@@ -158,7 +158,7 @@ public class RequestCallbackAdapter<T> implements RequestCallback {
   private final SerializationStreamFactory streamFactory;
 
   public RequestCallbackAdapter(SerializationStreamFactory streamFactory,
-      String methodName, long requestId, AsyncCallback<T> callback,
+      String methodName, int requestId, AsyncCallback<T> callback,
       ResponseReader responseReader) {
     assert (streamFactory != null);
     assert (callback != null);
